python/requirements.txt
matplotlib
pytest
numpy
pyusb
